Looks like it works now. Next, work on having AI players scale from little weenie enemies to ones with huge stats. This will mean reworking the DamageBacklog system, as well as some statuses.
